Inject dependencies to ScreenTextService.
Avoids the view pager on the visualizer sliding slightly after a pinch.

When pinching don't let the view become smaller than the original card width. Restore to the initial scale when pinch gesture is over.

Fixed unexpected click and scrolls when scaling the visualizer page.

Added a smoother way to resize the card view in the text visualizer.

Modify padding when resizing the page of the visualizer.

The card and text in the text visualizer are resized on pinch.