Updated to the latest version of Kotlin (1.9.23).
Also updated all the dependencies in the top-level gradle file, also had to update the navigation version in the app gradle file.

Updated Room and Dagger to their latest versions (2.6.1 and 2.50 respectively) to fix issues due to the new Kotlin version.
Also, enabled KSP because this version of Room supports it.