Fixed configuration change issue for the text info dialog for a word.
Fixed dictionary layout being empty.
Fixed definition and external links not updating when the language from changed (this was because the fragments in the adapter and the fragment manager were different so the listener was calling fragments not visible).
Fixed autocomplete AutoCompleteTextViews only having one item.
Fixed dialog disappearing in the process text activity.
Fixed changing the language to not working.

Also, unrelated to config changes:
Fixed links fragment not updating when changing to auto detect language.
Fixed the language label for autodetect not updating.